refactor(task-04): extract duplicated e2e helpers into shared module
Consolidate 8 helper functions (read_sessions, write_fake_orchestrator,
find_session, history_dir, read_tpm_config, setup_tpm_harness, plus
seed_tpm_plugin, seed_tpm_artifacts, list_dir_entries, events_jsonl_path,
config_dir) from 12 test files into tests/e2e/helpers.rs.

- Unify write_fake_orchestrator to always return PathBuf (superset)
- Unify setup_tpm_harness to return (Harness, TempDir, PathBuf)
- Replace find_session_by_title with find_session (identical logic)
- Replace read_sessions_json with read_sessions (identical logic)
- Fix duplicate mod tpm_prompt_injection in main.rs (merge artifact)
- Fix worktree_from.rs (outside touches_files but required for AC-05)

Net: -567 lines of duplicated code across 12 files.
