+++ tests/e2e/tui_config_settings.rs
@@ -8,55 +8,10 @@
 //! AC-05: Default `--tpm` (no extra flags) writes standard tier + TUI badge.
 
 use serial_test::serial;
-use std::path::Path;
-use std::process::Command;
 use std::time::Duration;
-use tempfile::TempDir;
 
 use crate::harness::{require_tmux, TuiTestHarness};
-
-// ---------------------------------------------------------------------------
-// Helpers (self-contained to avoid cross-file merge conflicts)
-// ---------------------------------------------------------------------------
-
-/// Drop a fake `agents/orchestrator.md` under `root`.
-fn write_fake_orchestrator(root: &Path) {
-    let agents = root.join("agents");
-    std::fs::create_dir_all(&agents).expect("create agents dir");
-    std::fs::write(agents.join("orchestrator.md"), "# Fake Orchestrator\n")
-        .expect("write orchestrator.md");
-}
-
-/// Create a harness with a git-initialized project and a fake plugin dir.
-fn setup_tpm_harness(name: &str) -> (TuiTestHarness, TempDir) {
-    let h = TuiTestHarness::new(name);
-    let project = h.project_path();
-
-    let git_init = Command::new("git")
-        .arg("init")
-        .arg("--quiet")
-        .arg(&project)
-        .output()
-        .expect("git init");
-    assert!(
-        git_init.status.success(),
-        "git init failed: {}",
-        String::from_utf8_lossy(&git_init.stderr)
-    );
-
-    let plugin_dir = TempDir::new().expect("plugin tempdir");
-    write_fake_orchestrator(plugin_dir.path());
-
-    (h, plugin_dir)
-}
-
-/// Read `.tpm/config.json` from the project directory inside the harness.
-fn read_tpm_config(h: &TuiTestHarness) -> serde_json::Value {
-    let path = h.project_path().join(".tpm/config.json");
-    let raw = std::fs::read_to_string(&path)
-        .unwrap_or_else(|e| panic!("failed to read {}: {}", path.display(), e));
-    serde_json::from_str(&raw).expect("invalid .tpm/config.json")
-}
+use crate::helpers::{read_tpm_config, setup_tpm_harness};
 
 // ---------------------------------------------------------------------------
 // AC-01 (Journey 11): Settings TUI → navigate to TPM category
@@ -117,7 +72,7 @@ fn settings_tui_shows_tpm_category_and_fields() {
 fn tpm_review_passes_seven_config_and_badge() {
     require_tmux!();
 
-    let (mut h, plugin_dir) = setup_tpm_harness("tui_cfg_review7");
+    let (mut h, plugin_dir, _) = setup_tpm_harness("tui_cfg_review7");
     let project = h.project_path();
 
     let output = h.run_cli_with_env(
@@ -176,7 +131,7 @@ fn tpm_review_passes_seven_config_and_badge() {
 fn tpm_disable_agents_config_and_badge() {
     require_tmux!();
 
-    let (mut h, plugin_dir) = setup_tpm_harness("tui_cfg_disable");
+    let (mut h, plugin_dir, _) = setup_tpm_harness("tui_cfg_disable");
     let project = h.project_path();
 
     let output = h.run_cli_with_env(
@@ -240,7 +195,7 @@ fn tpm_disable_agents_config_and_badge() {
 fn tpm_disable_implementer_stripped_and_badge() {
     require_tmux!();
 
-    let (mut h, plugin_dir) = setup_tpm_harness("tui_cfg_impl_prot");
+    let (mut h, plugin_dir, _) = setup_tpm_harness("tui_cfg_impl_prot");
     let project = h.project_path();
 
     let output = h.run_cli_with_env(
@@ -310,7 +265,7 @@ fn tpm_disable_implementer_stripped_and_badge() {
 fn tpm_default_flags_writes_standard_tier_and_badge() {
     require_tmux!();
 
-    let (mut h, plugin_dir) = setup_tpm_harness("tui_cfg_defaults");
+    let (mut h, plugin_dir, _) = setup_tpm_harness("tui_cfg_defaults");
     let project = h.project_path();
 
     let output = h.run_cli_with_env(
